+++ src/test/java/blackjack/domain/player/ParticipantResultTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import static java.util.Map.entry;
 import static org.assertj.core.api.Assertions.assertThat;
